interface: adding basic notifications
